LRCI-3932 Move job parameter definitions into its own package
