+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/job/definition/JobDefinitionFactory.java
@@ -32,6 +32,9 @@ public class JobDefinitionFactory {
 		else if (type == JobEntity.Type.PORTAL_PULL_REQUEST_SF) {
 			return new PortalPullRequestSFJobDefinition(type);
 		}
+		else if (type == JobEntity.Type.PORTAL_RELEASE) {
+			return new PortalReleaseJobDefinition(type);
+		}
 		else if (type == JobEntity.Type.PORTAL_UPSTREAM_ACCEPTANCE) {
 			return new PortalUpstreamAcceptanceJobDefinition(type);
 		}
