+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/job/definition/parameter/JenkinsBranchURLJobParameterDefinition.java
@@ -3,24 +3,24 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.job.definition;
+package com.liferay.jethr0.job.definition.parameter;
 
 import com.liferay.jethr0.util.StringUtil;
 
 /**
  * @author Michael Hashimoto
  */
-public class PortalPullRequestURLJobParameterDefinition
+public class JenkinsBranchURLJobParameterDefinition
 	extends BaseJobParameterDefinition {
 
 	@Override
 	public String getKey() {
-		return "portalPullRequestURL";
+		return "jenkinsBranchURL";
 	}
 
 	@Override
 	public String getLabel() {
-		return "Portal Pull Request URL";
+		return "Jenkins Branch URL";
 	}
 
 	@Override
@@ -36,12 +36,13 @@ public class PortalPullRequestURLJobParameterDefinition
 	@Override
 	public String getValueDescription() {
 		return StringUtil.combine(
-			"e.g. https://github.com/[user]/liferay-portal/pull/", "[number]");
+			"e.g. https://github.com/[user]/liferay-jenkins-ee/",
+			"tree/[branch]");
 	}
 
 	@Override
 	public String getValueRegex() {
-		return "https://github.com/[^/]+/liferay-portal(-ee)?/pull/[^/]+";
+		return "https://github.com/[^/]+/liferay-jenkins-ee/tree/[^/]+";
 	}
 
 }
