+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/job/definition/parameter/PortalBuildProfileJobParameterDefinition.java
@@ -3,22 +3,22 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.job.definition;
+package com.liferay.jethr0.job.definition.parameter;
 
 /**
  * @author Michael Hashimoto
  */
-public class PortalUpstreamBranchNameJobParameterDefinition
+public class PortalBuildProfileJobParameterDefinition
 	extends BaseJobParameterDefinition {
 
 	@Override
 	public String getKey() {
-		return "portalUpstreamBranchName";
+		return "portalBuildProfile";
 	}
 
 	@Override
 	public String getLabel() {
-		return "Portal Upstream Branch Name";
+		return "Portal Build Profile";
 	}
 
 	@Override
@@ -28,17 +28,17 @@ public class PortalUpstreamBranchNameJobParameterDefinition
 
 	@Override
 	public String getValueDefault() {
-		return null;
+		return "dxp";
 	}
 
 	@Override
 	public String getValueDescription() {
-		return "Insert your Portal Upstream Branch Name here";
+		return "Insert your Build Profile here (i.e. 'dxp' or 'portal')";
 	}
 
 	@Override
 	public String getValueRegex() {
-		return null;
+		return "(dxp|portal)";
 	}
 
 }
